--- modules/MMM-voice_control/liste_audio.py
@@ -1,11 +1,0 @@
-# import pyaudio
-
-# p = pyaudio.PyAudio()
-
-# for i in range(p.get_device_count()):
-#     print(p.get_device_info_by_index(i)['name'])
-
-
-import speech_recognition as sr
-for index, name in enumerate(sr.Microphone.list_microphone_names()):
-    print("Microphone with name \"{1}\" found for `Microphone(device_index={0})`".format(index, name))
